pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zenfolio"
version = "0.1.0"
description = "A minimal, powerful academic website generator built with ZenCFG"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["academic", "website", "static-site-generator", "zencfg"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "Topic :: Software Development :: Documentation",
]

dependencies = [
    "zencfg>=0.1.0",
    "markdown>=3.4.0",
    "jinja2>=3.1.0",
    "bibtexparser>=1.4.0",
    "python-frontmatter>=1.0.0",
    "nbformat>=5.0.0",
    "nbconvert>=6.5.0",
]

[project.optional-dependencies]
test = [
    "pytest>=6.0",
    "pytest-cov>=3.0",
]

[project.scripts]
zenfolio = "zenfolio:cli"

[project.urls]
Homepage = "https://github.com/JeanKossaifi/zenfolio"
Repository = "https://github.com/JeanKossaifi/zenfolio"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-v", "--tb=short"] 